+++ mfe-reference-data/src/app/features/coverage-types/coverage-type-list/coverage-type-list.component.ts
@@ -50,6 +50,9 @@ export class CoverageTypeListComponent {
   }
 
   delete(id:number){
+      const confirmed = window.confirm('¿Estás seguro de que deseas eliminar este tipo de póliza?');
+
+  if (!confirmed) return;
     this.errores.set([]);
     this.coverageService.deleteCoverageType(id).subscribe({
       next:(apiResponse)=>{
