+++ mfe-reference-data/src/app/features/regions/region-list/region-list.component.ts
@@ -42,6 +42,9 @@ export class RegionListComponent {
   }
 
   delete(id:number){
+      const confirmed = window.confirm('¿Estás seguro de que deseas eliminar este tipo de póliza?');
+
+  if (!confirmed) return;
     this.errores.set([]);
     this.regionService.deleteRegion(id).subscribe({
       next:()=>{
